+++ packages/layout/CHANGELOG.md
@@ -1,3 +1,57 @@
+## 0.2.18 (2026-08-18)
+
+### 🩹 Fixes
+
+- **layout,navigation-section:** use type intersection instead of interface extends for props ([335fc8f437c](https://github.com/migrationsverket/midas/commit/335fc8f437c))
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.21.1
+
+## 0.2.17 (2026-08-18)
+
+### 🩹 Fixes
+
+- **layout,sidebar:** widen style prop type to accept CSS custom properties ([2d038fffe8d](https://github.com/migrationsverket/midas/commit/2d038fffe8d))
+
+## 0.2.16 (2026-08-18)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.21.0
+
+## 0.2.15 (2026-08-17)
+
+### 🩹 Fixes
+
+- **layout,sidebar:** allow overriding width via CSS custom properties ([5cead0c5c7a](https://github.com/migrationsverket/midas/commit/5cead0c5c7a))
+
+## 0.2.14 (2026-08-14)
+
+### 🩹 Fixes
+
+- **components,badge:** pin badge to icon corner in width-constrained containers ([b1d8e302d10](https://github.com/migrationsverket/midas/commit/b1d8e302d10))
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.20.0
+
+## 0.2.13 (2026-08-13)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.19.0
+
+## 0.2.12 (2026-08-13)
+
+### 🩹 Fixes
+
+- **layout,navigation:** migrate off unresolvable @react-aria/collections dependency ([2322e21ef9a](https://github.com/migrationsverket/midas/commit/2322e21ef9a))
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.18.0
+
 ## 0.2.11 (2026-07-10)
 
 ### 🚀 Features